requirements.txt
langchain
openai
jinja2
colorlog
PyPDF2
pdfminer.six